Added if visible statements
Add if visible statements to the bar clock and clock windows to stop them from trying to make changes that didnt need to be made when the window wasn't in use. If their timers are going they will be doing nothing but checking if the window is visible each second.